requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
jinja2==3.1.2
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0
apscheduler==3.10.4
email-validator==2.1.0
aiosmtplib==3.0.1 